kiswahili-backend/requirements.txt
flask
flask-cors
gunicorn
openai
python-dotenv